Use shebang line to run pre-commit hook in interactive mode.
When the pre-commit hook is executed by a GUI tool, it will use a
non-interactive shell, which means that .bashrc, .profile, etc. will
be ignored. For many setups, this means that `npm` will not be on the
PATH.

Putting -i in the shebang line requests an interactive shell, so the
user can use .bashrc to ensure the PATH variable is set up correctly.