Capture and visualize trace for last iteration of the LR loop.
Ohm's left recursion works by repeatedly re-executing the body of a
left recursive rule, until an iteration fails to consume more input.
The last iteration of the loop contains important information that
should be shown in the visualizer -- i.e., why did it fail to make the
result bigger?

Previously, we discarded the trace from this last iteration, so it was
not possible to visualize it. With this commit, we retain the trace
information, and show it in the visualizer if the "Show failures" option
is enabled.