Show error message when overriding undeclared rule (fixes #150)
A recent refactoring (af18e7b) introduced a bug, and we were no longer
showing a good error message when a grammar would override or extend
an undeclared rule.

The problem was not detected due to a bug in the tests. This commit
fixes both the original issue and the problem with the tests.