BREAKING CHANGE: Eliminate default action for iteration nodes (#312)
Close #309 by removing the default semantic action for iteration nodes.

This means that for an iteration node n within an operation op(), calling n.op() will no longer have any default behaviour. Previously, the default action would be the equivalent of n.children.map(c => c.op()). Now, if a programmer wants this behaviour, they need to implement it manually — either by writing it in the individual semantic actions, or by writing an _iter action that does this for all iteration nodes.

This PR also reverts the change made in #308. Eliminating the default action for iteration nodes solves the same problem that #308 was addressing. It's not necessary to do both, and it's simpler if the default action applies to non-terminal nodes with a single child rather than non-terminal nodes with a single child which is not an iteration node.