Fix #363 by removing imports of ohm-js from extras (it’s not required) (#366)
The problem described in #363 occurs because the import of `ohm-js` pulls in the ES module, while `ohm-js/extras` is a CommonJS module. (This is the [dual-package hazard](https://nodejs.org/api/packages.html#dual-package-hazard).)

This PR eliminates the import of `ohm-js` from the extras module. It's not really required: it was only used for `instanceof` checks and we can skip those and instead check that the methods we need actually exist.